Reorganize card tests by expansion
