+++ tests/test_plunder_card.py
@@ -2,7 +2,26 @@ from dominion.cards.registry import get_card
 from dominion.game.game_state import GameState
 from dominion.game.player_state import PlayerState
 
-from tests.utils import DummyAI
+from tests.utils import ChooseFirstActionAI, DummyAI
+
+
+class StopAfterOneVillageAI(ChooseFirstActionAI):
+    """AI that stops First Mate after playing a single copy of the named card."""
+
+    def __init__(self):
+        super().__init__()
+        self._first_mate_resolution_calls = 0
+
+    def choose_action(self, state, choices):
+        non_none = [c for c in choices if c is not None]
+        if non_none and all(card.name == "Village" for card in non_none):
+            if self._first_mate_resolution_calls == 0:
+                self._first_mate_resolution_calls += 1
+                return non_none[0]
+            return None
+
+        self._first_mate_resolution_calls = 0
+        return super().choose_action(state, choices)
 
 
 class InsigniaChoiceAI(DummyAI):
@@ -126,3 +145,43 @@ def test_insignia_respects_optional_topdecking():
     assert player.deck[1].name == "Silver"
     assert ai.decisions == []
     assert ai.seen_cards == ["Silver", "Estate", "Gold"]
+
+
+def test_first_mate_effect():
+    ai = ChooseFirstActionAI()
+    state = GameState(players=[])
+    state.initialize_game([ai], [get_card("First Mate"), get_card("Village")])
+
+    player = state.players[0]
+
+    player.hand = [get_card("First Mate"), get_card("Village"), get_card("Village")]
+    player.deck = [get_card("Copper") for _ in range(10)]
+    player.discard = []
+    player.actions = 1
+
+    state.phase = "action"
+    state.handle_action_phase()
+
+    assert sum(1 for c in player.in_play if c.name == "Village") == 2
+    assert any(c.name == "First Mate" for c in player.in_play)
+    assert len(player.hand) == 6
+
+
+def test_first_mate_can_stop_after_first_copy():
+    ai = StopAfterOneVillageAI()
+    state = GameState(players=[])
+    state.initialize_game([ai], [get_card("First Mate"), get_card("Village")])
+
+    player = state.players[0]
+
+    player.hand = [get_card("First Mate"), get_card("Village"), get_card("Village")]
+    player.deck = [get_card("Copper") for _ in range(10)]
+    player.discard = []
+    player.actions = 1
+
+    state.phase = "action"
+    state.handle_action_phase()
+
+    assert sum(1 for c in player.in_play if c.name == "Village") == 1
+    assert any(c.name == "Village" for c in player.hand)
+    assert len(player.hand) == 6
